remove old config helpers, rename new ones to get rid of New prefixes
